Re-factor Slope and Aspect map ops in preparation for supporting new args for aspect
 - for issue #352
